Restructure log code to be its own feature
