+++ src/app/features/editor/panels/RuntimePanel/RuntimePanel.tsx
@@ -17,8 +17,8 @@ import { useActions } from "../../../../../lib/useActions";
 import { editorActions } from "../../actions";
 import { PendingGameRenderer } from "../../../compiler/GameRenderer";
 import type { PanelProps } from "../definition";
-import { colors } from "../../components/Log/colors";
-import { LogIdentifier } from "../../components/Log/types";
+import { colors } from "../../../log/colors";
+import { LogIdentifier } from "../../../log/types";
 import { RuntimeErrorFallback } from "./RuntimeErrorFallback";
 import { CompilingIndicator } from "./CompilingIndicator";
 import { useEditorGameCompiler } from "./useEditorGameCompiler";
