+++ src/app/features/editor/slice.ts
@@ -19,6 +19,7 @@ import {
   addNodeBySplitting,
   removeNodeByKey,
 } from "../../../lib/reactMosaicExtensions";
+import type { LogEntry } from "../log/types";
 import type {
   EditorObjectId,
   EditorState,
@@ -31,7 +32,6 @@ import { selectors } from "./selectors";
 import { moveObject } from "./reducers/moveObject";
 import { panelStorage, selectedObjectStorage } from "./localStorage";
 import { createPostProcessReducer } from "./reducers/createPostProcessReducer";
-import type { LogEntry } from "./components/Log/types";
 
 const initialState: EditorState = {
   syncState: "dirty",
