+++ src/app/features/log/components/SpreadLogValue.tsx
@@ -1,4 +1,4 @@
-import type { LogValue } from "../../types";
+import type { LogValue } from "../types";
 import { DynamicLogValue } from "./DynamicLogValue";
 
 export function SpreadLogValue({ parts }: { parts: LogValue[] }) {
