+++ src/app/features/log/isPrimitive.ts
